requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
fastapi-cdn-host==0.1.0
transformers==4.35.2
pillow==10.1.0